+++ releasenotes.rst
@@ -1,11 +1,13 @@
 Kiwi Release Notes
 ==================
 
-Wrappers x.x.x | Solver x.x.x | unreleased
+Wrappers 1.0.2 | Solver 1.0.1 | unreleased
 ------------------------------------------
 - prevent attempting a dual optimize on a dummy row PR #56 closes #15
 - add ``dump`` and ``dumps`` methods to inspect the internal state of the
   solver PR #56
+- test on Python 3.7 PR #51
+- improvemnts to setup.py and tests PR #46 #50
 
 Wrappers 1.0.1 | Solver 1.0.0 | 24/10/2017
 ------------------------------------------